+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Fixing issue with response single choice and bipolar likert choice where keyboard navigation was not working as expected in table mode.
+- Center text and number inputs content when used in a cloze question.
 
 ## 1.9.0 - 2022-02-04